Revert "Code refactoring"
