+++ .gitignore
@@ -1,11 +1,4 @@
-# Ignore user-specific config
-vp/configs/user_config.py
-
-# Ignore cookies directory
-vp/crawling/cookies/
-
-# Ignore virtual env
 venv/
 
-*__pycache__
 video_preprocessor.egg-info/
+__pycache__/
